requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6
cffi==1.17.1
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.3
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-js-asset==3.1.2
django-mptt==0.17.0
django_roles==0.0.4
djangorestframework==3.14.0
kombu==5.5.3
mysqlclient==2.2.7
pillow==11.2.1
prompt_toolkit==3.0.51
pycparser==2.22
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
redis==5.0.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
